+++ packages/@storefront/sayt/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] [patch]
+### Changed
+- SF-1279: Add `isHovered` boolean state to `gb-sayt-autocomplete` and toggle state property when activation is set.
+- SF-1279: Add `isActiveAndHovered` method to `gb-sayt-autocomplete`.
+- SF-1279: Use a selector to access the config in `gb-sayt-autocomplete` `onHover`.
+
 ## [2.5.0] - 2019-02-04
 ### Changed
 - Update `@storefront/core` to 2.5.0.